Increase margins for box skewness checks
The margins for checking for box skewness were to tight for
infrequent pairlist updated and strong pressure coupling or
box deformation.

Also contains some cleanup of the code concerned.

+++ api/legacy/include/gromacs/pbcutil/pbc.h
@@ -187,7 +187,7 @@ PbcType guessPbcType(const matrix box);
  * \param[in] box   The simulation cell
  * \return TRUE when the box was corrected.
  */
-bool correct_box(FILE* fplog, int step, tensor box);
+bool correct_box(FILE* fplog, int64_t step, tensor box);
 
 /*! \brief Initiate the periodic boundary condition algorithms.
  *
